+++ include/slug_ffi_prototype.h
@@ -16,16 +16,21 @@ typedef enum {
 typedef struct slug_ffi_host_api slug_ffi_host_api;
 typedef struct slug_ffi_call slug_ffi_call;
 
+typedef struct {
+  const char *data;
+  uint64_t length;
+} slug_ffi_text;
+
 typedef bool (*slug_ffi_argument_i64_fn)(slug_ffi_call *, size_t, int64_t *);
 typedef bool (*slug_ffi_argument_f64_fn)(slug_ffi_call *, size_t, double *);
 typedef void (*slug_ffi_set_i64_fn)(slug_ffi_call *, int64_t);
 typedef void (*slug_ffi_set_f64_fn)(slug_ffi_call *, double);
-typedef void (*slug_ffi_set_error_fn)(slug_ffi_call *, const char *, const char *);
+typedef void (*slug_ffi_set_error_fn)(slug_ffi_call *, slug_ffi_text, slug_ffi_text);
 
 struct slug_ffi_host_api {
   uint32_t abi_major;
   uint32_t abi_minor;
-  size_t table_size;
+  uint32_t table_size;
   slug_ffi_argument_i64_fn argument_i64;
   slug_ffi_argument_f64_fn argument_f64;
   slug_ffi_set_i64_fn set_i64;
@@ -33,22 +38,24 @@ struct slug_ffi_host_api {
   slug_ffi_set_error_fn set_error;
 };
 
-typedef slug_ffi_status (*slug_ffi_callback)(const slug_ffi_host_api *, slug_ffi_call *);
+typedef int32_t (*slug_ffi_callback)(const slug_ffi_host_api *, slug_ffi_call *);
 
 typedef struct {
-  const char *name;
-  size_t minimum_arity;
-  size_t maximum_arity;
+  uint32_t descriptor_size;
+  slug_ffi_text name;
+  slug_ffi_text member_key;
+  uint64_t minimum_arity;
+  uint64_t maximum_arity;
   slug_ffi_callback callback;
 } slug_ffi_function_descriptor;
 
 typedef struct {
   uint32_t abi_major;
   uint32_t abi_minor;
-  size_t descriptor_size;
-  const char *module_name;
+  uint32_t descriptor_size;
+  slug_ffi_text module_name;
   const slug_ffi_function_descriptor *functions;
-  size_t function_count;
+  uint64_t function_count;
 } slug_ffi_module_descriptor;
 
 typedef const slug_ffi_module_descriptor *(*slug_ffi_module_init_fn)(const slug_ffi_host_api *);